+++ src/AppBundle/Form/Type/WLT/ProjectType.php
@@ -113,6 +113,15 @@ class ProjectType extends AbstractType
                 'placeholder' => 'form.no_template',
                 'required' => false
             ])
+            ->add('weeklyActivityReportTemplate', EntityType::class, [
+                'label' => 'form.weekly_activity_report_template',
+                'class' => ReportTemplate::class,
+                'choice_label' => 'description',
+                'choice_translation_domain' => false,
+                'choices' => $templates,
+                'placeholder' => 'form.no_template',
+                'required' => false
+            ])
             ->add('studentSurvey', EntityType::class, [
                 'label' => 'form.student_survey',
                 'class' => Survey::class,